Change background map, player, bullet and wall sprite
In order to investigate how TextureWorld is used on the original code
based of "Shadow", changing the background map to suit the theme of the
game (Ship based game) was necessary.

In the original code base, walls always have no texture and identified
to be black. This black walls does not suit the theme of the game. Since
the ShadowFactory detects ShadowCaster by its own, and have
CollisionWithWorld method, the code can stay as it is right now to stop
player from going through wall/obstacle. In the TexturedWord, Load()
function have detection of the pixels on the map and decided what
texture to be set on top of the current pixels. sf::Color::Black check
is added in order to set black pixels to have Wood tiles texture, hence
giving the walls/obstacle a wooden texture while still having collision
checks with player

Potential issue will arise if Shadow is never used outside of collision
detection will player (act as wall) then there is waste of codes that
needed to be refactor