Create TestPacket scenario from client and server
In order to create a lobby mode, TestPacket is created to mockup and
investigate how the code base process incoming packets and sends out
packets both from client and server side.

RoboCatClient have an extra input for "Enter" key that will change
NetworkManagerClient state into NCS_Test, and sends a TestPacket to
server. Server will add the player ID to a map mPlayerReady and checks
if player already exist in that map. Once mPlayerReady size is equal to
mPlayerToClientMap, server will send a packets to all client.

This TestPacket is a foundation for a lobby state design where the game
will start if all players hit ready button (enter key).