Create pseudo lobby state
Pseudo lobby state is where all players unable to move or shoot until
all players have pressed Enter key. This allows for people to join in
first then start at the same time. There is no player join restriction
once the game is on running state, but it is possible to modify the
current iteration to handle player join