Create Ready UI and Refactor SpawnPoint
Create a ready UI text on top right corner of the game. The text and
color will change depend if a player is ready or not. The ready UI will
not be rendered when its out of lobby state

Refactor SpawnPoint to prevent more than 2 player spawn in the same
place. To do this, created a struct SpawnPoint that holds position and
playerId. Occupied SpawnPoint will hold the playerId, and the unoccupied
will hold -1 for playerId value.